refactor(cli): move onboarding web search and build config
